Rework the comment service to it adapts to the (new) HTTP GW
Add the roomId for every comment related request so the GW can get room access info.

Rework the comment resolver to first get the room, then the comment.
This is needed for getting the roomId and adhering to the GW structure for now.
ShortId -> roomId resolvment could be moved to the GW in the future.