Install Roboto font package instead of using Google Fonts
As a side-effect this changes displayed font-weight because fonts for
some weights were missing before. In some cases weight values now need
to be corrected.

Material by default uses font-weight 500 for bold display which was
incorrectly displayed like regular text (400). We on the other hand explicitly
set the weight to 'bold' (700) in some cases which now leads to
inconsistent font rendering.