Refactor authentication handling
`ClientAuthentication` is now used throught the app and replaces the
`User` class. Previously, it was only used to handle the API response.
`User` was basically the same as `ClientAuthentication` with a few
additional properties which became obsolete properties which became obsolete.

`AuthenticationService`'s `getUser`, `getUserAsSubject` and `watchUser`
methods have been replace with a single `getAuthentication` method which
returns an `Observable<ClientAuthentication>`. Internally, a
higher-order `Observable` is used to emit changes to authentication as
a stream.

The logic of `AuthenticationService`'s login-related methods have been
refactored. Instead of returning the strings `true`, `false` and
`activation`, a `ClientAuthenticationResult` is return which contains
the status and the auth object if successful.

Known issue:
* Role-based, conditional rendering for elements of the header bar no
  longer works.