Add the ability to transfer rooms from a guest account
Rooms from a local guest account can now be loaded from the user's home
view. The controls are only shown if the current account is not a guest
account and a token for a guest account is stored. The list is loaded
without interaction if the user doesn't have any memberships yet.
The button for room transfers is placed in the search-bar in an 'more'-menu.
The searchbar will always be displayed now if there are rooms.
Otherwise the transferable rooms will be displayed directly.

Co-Authored-By: Lukas Mauß <lukas.mauss@gmx.de>