Always join rooms as participant
When joining a room by entering the short ID, the view for participants
is always used instead of navigating based on the role.

The old behavior didn't work as expected, because the check for room
ownership was incorrect. As a consequence, moderators got the moderator
view while owners got the participant view.