Improve room join and history handling
Handling of auto login and adding of room history entries has been
removed from components (`RoomJoinComponent`,
`RoomParticipantPageComponent`. The auth guard is now also active for
participant role and handles those things.

As a result, it now possible to enter the app using any route for
participants w/o prior authentication. Furthermore, this fixes duplicate
requests to retrieve the room and update the history. The room history is
now only updated if the user doesn't already have a membership.

This also adds an event fired when joining a room for the first time to
reset locally cached memberships. The room is now shown in the list w/o
reloading the app (this was a regression introduces by !624).

Closes: #623
Related: !624, !644, !647