Subscribe to the rooms WS message stream when joining
The subscription allows the backend to keep track of active users in a
room. Furthermore, is can be used in the future to react on room-wide
events.

The subscription is currently triggered by the RoomResolver. While this
logic would be a better fit for the guard the resolver has the room's ID
which is required for the subscription.