Add RenderedTextComponent which uses server-side rendering
The RenderedTextComponent can either use an already rendered text value
or dynamically query the server when the raw text value changes.

Additional fields for rendered texts have been added to models.

Usage of the component to render statically (already rendered properties
in the model):

    <app-rendered-text
        [rawTest]="# My Heading\n\nMarkdown formatted text"
        [renderedText]="<h1>My Heading</h1><p>Mardown formatted text</p>"
    </app-rendered-text>

Usage for the component to render dynamically:

    <app-rendered-text
        [rawTest]="# My Heading\n\nMarkdown formatted text"
        [dynamic]="true"
        [markdown]="true"
        [latex]="true"
        [linebreaks]="true"
        [markdownFeatureset]="markdownFeatureset">
    </app-rendered-text>