Migrate to Chart.js 3
Chart.js 3 is packaged as ES modules, so only the components which are
explicitly imported will be part of the bundle. Futhermore, Chart.js 3
does no longer depend on Moment.js which can then be removed later.
Version 3.0 of Chart.js is still in beta but should be stable enough for
production, especially since only basic functionality is used at the
moment.

Related: #626