Navigate to login page if anonymous room creation is disabled
Known issue:
The dialogs background overlay flickers during navigation because the
dialog is opened and then closed immediately. To fix this issue, the
auth logic would need to be moved to the dialog service or even the
component which opens the dialog (ugly) or dialog handling in general
would need to be refactored.

This doesn't really have a large impact on the user experience, so I
don't think it needs fixing at the moment.