Improve API config handling
A resolver is now used to make the API config available for components.
Furthermore, the config is no longer cached indefinitely but expires
after one hour. Dialog components currently still need to retrieve the
config from the `ApiConfigService` explicitly because they are not
routed and so cannot use the resolver.