Adjust handling of app updates based on a new importance property
There are three importance levels:
* OPTIONAL: The update is performed silently when the user visits the
  next time or reloads manually. This is intended for minor improvements
  and bugfixes.
* RECOMMENDED: The update dialag is shown after the service worker has
  loaded the newer version (previous behavior).
* MANDATORY: The update dialog is shown immediately. The update button
  is disabled util the service worker has loaded the newer version. This
  is intended for breaking changes in the API.

Because the user might have skipped previous updates, the actual
importance level is determined based on multiple versions. So instead of
the latest version, a list of recent versions is now retrieved. The
actual importance is the highest importance of all versions newer than
the currently installed one. If the installed version is no longer part
of the list, `MANDATORY` is implied.