Improve mandatory update handling
The ID of the latest available version is now stored locally. If app is
loaded at a later time, the importance of an `MANDATORY` update is
degrated to `RECOMMENDED` when its ID is lower or equal to the latest
locally stored ID. As a result, the update dialog is only forced once.
This is done to prevent blocking of the UI permanently if something went
wrong.

Instead of a single string for the version ID an object with version
info is now stored in localStorage with the following properties:
* installed (ID of installed version)
* latest (ID of latest available version)
* hash (Hash of the installed version from NGSW, not used yet,
  intended for debugging/tracking)