Update URL before tracking events
The URL is now updated for tracked events. For HTTP error events the URL
of the failed request is used instead of the client route. The HTTP
status code is now set as event name instead of setting it as value.
The unreliable statusText is no longer used. It falls back to `OK` which
doesn't make sense for errors.  Furthermore, the room ID is now always
replaced with a placeholder.