Adapt components to model changes
Use the new comment and group state for creator and participant components. Contents now can be published as block of many sequential contents or as single content.
The sorting of the contents can change the publish too: If a not yet published content will be moved into a group of published contents, the content will be published too after saving.
Otherwise, if a content in a block of published contents will be moved outside this block, the content will be locked.