Switch to fontsource package for text font
The previously used typeface project is now deprecated and will no
longer be updated. Font-related (S)CSS and imports have been moved to a
separate SCSS file.

As a side effect, this also adds webfont support for non-latin charsets.
Previously, a fallback system font was used in this case. A separate
font file is used and lazy loaded for the unicode range of the charset.