Remove stale cache entries
The cache is now cleared after WebSocket reconnections. If there are
connection issues the cached items are still available but once the
connection is reestablished, they will be removed.

A grace period of 10 seconds is given. If the connection is
reestablished in this period, the cache will not be cleared. This is
done to avoid punishing bad connections which are particularly dependend
on the cache.

Futhermore, the cache is now explicitly pruned in intervals so the
dispose handlers are called early and subscriptions for old cache
entries can be cancelled.