Add notification for published contents in nav bar
Show notification in participants nav bar when creator publish new contents to a content group which is published.
If the creator publishs contents to a group which is not the active one on the participant side, the notification will appear too and the route to the group will be changed.
The default group is the first group with published contents now.