Improve URI state handling for participants' content carousel
The URI state is now also updated when navigating to the first content.
Previously, the index was only set as part of the URI starting with the
second content. Since the URI without an index automatically navigates
to the first unanswered content, reloading the page for the first
content could previously trigger an unwanted navigation.

Furthermore, the URI has been fixed so that an additional encoded slash
is no longer added (e.g. /group/mygroup/%2F2).

Related: !959