Allow multiple answer rounds
Use extensions for starting new round for a content. For this restructure the context menu for contents in the content list view. Also embed the extension for the selection of which rounds the statistics should be display.
Furthermore refactor content choice statistics so that multiple rounds can be displayed. At least show statistics of current round in statistic list and add new column to display current round.