Rework likert scale format to support multiple option templates
The format now uses client-side templates for answer options. Instead of
the individual options, the selected template is stored for the content.

The following templates have been implemented:
* Agreement (Strongly agree <-> Strongly disagree)
* Intensity (Extremely <-> Not at all)
* Frequency (Always <-> Never)
* Quality (Very good <-> Very poor)
* Probability (Definitely <-> Definetly not)

All templates use a five-point scale by default. The neutral option can
be disabled. In this case the middle option is removed, leading to a
four-point scale.