Add caching support for non-entity endpoints
A new `AbstractCachingHttpService` has been added which allows caching
of the responses of endpoints which are not related to a specific
entity.

The new inheritence hierarchy now is:

    AbstractEntityService -> AbstractCachingHttpService -> AbstractHttpService

In contrast to the `AbstractEntityService`, the new service supports
caching of items but cannot update them out-of-the-box. Custom logic for
the endpoint is needed to react to the appropriate WS change events.

The `CachingService` has been refactored to support separate caches.
This now used to have one shared cache and a separate cache for the
current room. This has the advantage that the full room cache can be
cleared on room changes while items of the shared cache can be kept.