Prevent inserting text when shortId length limit has reached
A separate handler for `beforeinput` events is used to prevent the
default action when the length limit has been reached. In contrast to
the `ìnput` event, its specification is quite new and only supported by
current browser versions. This is a minor UX enhancement, so the field
is still fully functional if the handler cannot be used.