Remove obsolete enableModeration flag and related code
Disabling moderation does not have any benefit for users so the setting
is now completely removed. Already before these changes, the flag was no
longer used, to hide the moderation feature.

This also fixes a bug that an endless loading indicator for the comment
list was shown if the flag or extensions object were not present.