Fix and simplify extension initialization for comment settings
The component did not reliably initialized the extensions and its nested
objects. As a consquence, tags were not correcly loaded if no other
extensions setting were changed (e.g. through enabling feedback or
changing the comment score threshold).

The extension and nested objects are now always initialized if they are
not already set. This also allows the use of simplified conditions.

With the changes of !1103 this issue happens in more cases because it is
more likely that no extension settings exist but the bug already existed
before.

Related: !1103