Use an outlet for the sub routes of the admin area
With the use of an outlet, the routing is now more consistent with how
routing is implemented for the rest of the application. It is no longer
necessary to pass the route as data and use custom routing logic for sub
routes. Components are no longer hard wired inside the
`AdminHomeComponent`. As a noticible effect, the summary bar is no
longer reloaded when switching sub routes.

Closes: #782