Add new `SeriesResultsComponent` which can be used as overview after answering contents
The component shows an overview of the contents answered. If there are contents with correct answers a total score is shown. Otherwise a clipart is used at this place.

Also routing and content group service has been extended for needs of new component. A new model `AnserResultOverview` has been introduced.