Send 'Cache-Control' header for immutable assets
Instruct browsers to cache assets with a hash as part of the file name
for up to 6 month and to do not send any requests for them, even when
reloads are performed.

`immutable` is an extension supported by Firefox and Safari. Chrome does
not support it but uses heuristics to apply a similar behavior.