Add smart pausing to comment list
Comment stream is now paused automatically while scrolling. If there are new comments, a button appears which brings user back to the top.

When user is scrolling to the top, the new comments are displayed.

The pause/play button is kept for now, since it could be reused for a going into a readonly state.