Fix styles which are broken by switching to MDC
This also fixes some issues with the dark theme since this is now declared as dark theme and a lot of custom styles are no longer necessary.

Also now don't force overriding of all styles customized in the past. This affects following components:

1. Form Field
Stop using custom border color on focus which leads to worse UX and a11y. Default for this is now used which is the primary color.

2. Room join input field
Instead of using a custom, very oversized input for that, just make the input a little bit higher and use no custom sized icon button. At this point switch to a better arrow icon which matches in the button perfectly.

3. Snackbars
Use default background color for snackbars which offers a lot better contrast to background and surface and will get more attention.