Migrate to modern Yarn version
Starting with version 2 Yarn uses a new lockfile format and changes how
modules are resolved by Node.js when the PnP linker is used.

If the PnP linker should cause any compatibility issues in the future,
we can simply switch back to using a `node_modules` folder by setting
`nodeLinker: "node-modules"` in `.yarnrc.yml`.

Furthermore, the package name has been changed to arsnova-webclient to
be consistent with the repository/project name. The version property has
been removed since it was not maintained and is not useful for a private
package which is not published.

Closes: #1058
See: https://yarnpkg.com/features/pnp