Set fix timeout before automatically navigate to next content
The condition to determine the timeout before navigating to next content after answering is broken since `correctOptionsPublished` has been moved to content group.

Since the participant gets a feedback no matter which format or if answers are published, the timeout should be consistent and always 1000ms so that the user has enough time to take notice of the feedback.