Show correct combination in sort statistic chart no matter how often it was selected
To be able to get the answer count or percentage of participants answered with the correct sort, it is necessary to display the combination in chart no matter if other cominations have been selected more often.

The correct option is then moved to the position before the sum of other combinations.