Fix expired auth token refresh
The logic for refreshing auth tokens did not subscribe to the outer
`Observable` which handles errors. This might have been intentionally
(maybe to delay the redirect until the next user interaction) because
the `Observable` for the actual request is turned hot (`shareReplay`),
so request were going out and the auth state was updated on success. But
the missing error handling caused an invalid application state which
required reloading to fix when an expired token was refreshed.

Closes: #1139