Add caching support for multiple entities
Caching is a little more involved when multiple entities are retrieved
at once. Some but not all entities might already be cached. In this
case, only the missing entities are retrieved while the cached entities
are transparently added to the list. The entities are sorted
corresponding to the original order of IDs.

Closes: #1141