Allow caching for entities w/o subscription
Previously, entities could not be cached w/o individual subscription
because they would get stale when there are server-side changes to them.
Now, the room-wide subscription is used to detect when entities become
stale and remove them from the cache if the is no individual
subscription which would lead to a cache update.

`AbstractEntityService`s `useSharedCache` has been renamed to
`useChangeSubscriptions` because it is now also used to control the
subscription behavior. Both share the same state because without
individual subscriptions, the shared cache cannot be used.

Closes: #1143