Setup Storybook for project
Storybook can be started with `ng run arsnova-webclient:storybook`.

For getting Storybook to run `node-modules` has to be used for `nodeLinker` in
yarn config.

With `./storybook/preview-body.html` and `.storybook/style-loader.scss` the
theming is fixed for testing with stories. This is needed because there is no
option to use the `angular-material-css-vars`-lib from here.