Create service, directive and components for loading forms
To indicate that a form is currently loading while performing a request, input elements should be disabled and action buttons should indicate loading.

For this create the following:

`FormService`: Handle and provide current form state (disabled or enabled).

`DisableFormDirective`: Disable input fields and other elements while form is disabled via service. Use via `appDisableForm`.  This directive can be used if the component doesn't extend to `FormComponent` where `formDisabled` can be used for disabling elements.

`FormComponent`: Abstract component for simply bringing disabling/enabling of current form to other components. Communicates with form service and provides `formDisabled` variable, which is updated automatically. If a form control is set via `setFormControl(ormControl: FormControl)`, it will be disabed/enabled automatically according to current state as well.

`LoadingButtonComponent`: Component which displays a `mat-raised-button` or `mat-flat-button` according to `isDialog` flag. While the current form is disabled after button is clicked, the button is disabled and a loading indicator is displayed instead of the button text.