Indicate loading and disable form for components of creator, presentation and shared module
Use `LoadingButtonComponent` and disable form inputs while loading for the following:

Creator:
* Series creation dialog
* Series deletion dialog
* Series answer deletion dialog
* Content creation/edit
* Content deletion dialog
* Content answer deletion dialog
* Start new round dialog
* Comment list deletion dialog
* Comment deletion dialog
* Comment answering
* Room deletion dialog
* Room settings edit
* Edit access rights
* Remove access rights dialog
* Create/Edit announcement
* Delete announcement dialog
* Start/Stop live feedback
* Change live feedback type

Presentation:
* Publish series dialog

Shared:
* Comment creation dialog
* Content answer deletion/ban dialog
* Formatting toolbar

For this, refactor live feedback logic for creator and presentation page so that loading button and other components can wait for success of starting/stopping feedback or changing type.

Also refactor room settings. Use the loading button here as well and stop disabling it if there are no changes for consistency.