Update translation files for new scopes
Duplicate l10n string which are used in multiple modules have been moved
to a shared translation file. In some cases there were inconsistent
translations for the same key. The selected translation is based on the
order in which translations were processed which might not have been the
best solution for each individual string. Therefore, the strings should
be manually reviewed at a later time.