Refactor `BaseDialogComponent` for l10n
Stop using `section` as a scope for input strings and just use the variables instead.

Remove unnecessary aria labels for confirm and cancel buttons which are currently broken if the component isn't used for deletion.

Return `undefined` if dialog action has been cancelled to simplify result checks in `afterClosed` subscriptions.