Create reusable components for template tag and language selection
Create `TemplateTagSelectionComponent` for selection of template tags. Selected tags were emitted via `selectedTagsChanged` and can be set initially with `selectedTags` input. Creation of new tags on the fly is disabled by default and can be enabled with `allowCreation`.

Create `TemplateLanguageSelectionComponent` for selection of template lang. Selected lang is emitted via `selectedLangChanged` and can be set initially with `selectedLang`.

Use these components in `ContentGroupTemplateEditingComponent`.