Create content group template selection and add to creators room overview
Create `ContentGroupTemplateComponent` which displays all relevant informations of content group templates as a list item. Selection event is emitted via `templateSelected`.

Create `ContentGroupTemplateSelectionComponent` which displays public and own templates in a dialog. For filtering public templates, language and tags can be selected. If a template selection event is received, a copy of the template is added to the room.

Add new button "Use Template" to creators room overview which opens the content group template selection dialog.