Create preview for content group template selection
Add new preview button next to add button in content group template selection. In preview show full description, all tags, link to license and list of content templates.

For this add new `ContentGroupTemplatePreviewComponent`.

Also create `TemplateLicenseComponent` for displaying the license with and without link and use it in `ContentGroupTemplateComponent` as well. For this extend LICENSES map.