Update paths in SCSS to library assets
The use of `~` to reference library assets was deprecated by Webpack and
is not supported by esbuild. Library folders will be implicitly
included.

Webpack for now is still used with Karma and Storybook and seems to have
an issue with non-(s)css library assets referenced without `~`. As a
workaround, a separate `fonts.scss` is used for them.