Avoid showing cookie consent dialog if not needed
The statistics cookie category is now disabled if:
  * DNT is enabled in the browser ("Do Not Track")
  * no tracking URL is defined in the UI config

Handling of consent settings has been adjusted to handle the case that
categories are later added or removed. Categories which are disabled or
have a default consent value do not require user consent. If there are
no categories which require consent, the dialog is now be skipped.

Note: DNT has been deprecated since 2018 due to lack of adoption. But
major browsers still support it and users might expect to not be tracked
when they use this browser setting. At this time, still around 80 % of
users use a browser with DNT support. The code has no effect on browsers
which do no longer support DNT. We can remove it once no major browsers
support it anymore.