Fallback to cookieless tracking
Cookieless tracking is now used if no consent is given for tracking
cookies. If a user does not want to be tracked at all, they can signal
this by using the browser's "Do Not Track" setting in which case
tracking will be completely disabled.

For Matomo this also has the effect, that already created tracking
cookies are deleted when statistical cookies are disabled later.
Previously, cookie settings only effected cookies which where created by
the app itself.