Implement answer statistics for numeric format
Vizualize answer results in a custom bar chart:
* Answers are grouped according to range size with group sizes of factors of 1,2 and 5
* Answer labels are formatted to compact localized number if greater than 10.000
* Bar tooltips display group range

Below the chart an info section is displayed with contains mean, median, standard deviation, variance, minimum and maximum for current round(s).

For displaying the correct answer (if existing) the info section is replaced with displaying of the correct number (and tolerance if existing) and how many participants have answered correctly.