Restructure and redesign comment list bar
To ensure that the comment list bar is as consistent as possible and to highlight the most important features depending on the role view, make the following changes:

General:
* Use full width for both, static and fixed bar
* Use consistent display of fixed bar when scrolling
* Always show search input instead of icon
* Split filtering and sorting and show current sorting with an select input on desktop devices
* Move time period selection to filter menu
* Add selection for categories to filter menu
* Stop using `mat-mini-fab` that we don't use anywhere else
* Use more common icon for filter button
* Show active filter (period and filter or category) as chips below bar
* Improve wording for some of the list bar actions (de/en)

Participants:
* Use `mat-raised-button` for the primary action (post creation) on desktop devices

Creator:
* Always show tabs for switching between public and moderated posts and give them an own column
* Move button for creating new posts to more menu
* Show button for post creation if there are no posts for moderation as well
* Use more icon for consistency in more menu trigger button

These changes shouldn't affect any logic, though some changes had to be made.

No longer needed `MenuDividerComponent` has been removed.

To ensure a correct calculation of the fixed bar width, a new util service, `UiService`, has been introduced with which the scrollbar width for the device/browser can be determinated and set as css variable `--scrollbar-width`. This is done by a call in `AppComponent`.