Cleanup Storybook dev dependencies
While most of the transitive dependencies are still used, they no longer
need to be declared explicitly. Those, which still need to be declared
have been moved to Yarn's packageExtensions.

This reduces the number of commits and CI builds created for updates of
transitive development dependencies which we do not really care about.