Add server time service for time synchronization
A new interceptor has been introduced which
calculates the current server time by measuring
the latency of requests, adding it to the time
from the Date header received in the response.

A problem with this approach is that the header
does not contain milliseconds. But the inaccuracy
is compensated by averaging the values of all
requests.

The average offset between local and server time
is managed by the server time service, which uses
it to calculate the server time at any time.

Closes: #1271