Add gamification elements to participants content answering
If quiz mode is enabled for a content group, in participants content answering `ContentWaitingComponent`, `CountdownTimerComponent` and `ContentLeaderboardComponent` are used if needed.

Also participants are routet to new published contents if quiz mode is enabled. If answering of a content is locked due to finished timer, the answer input will be disabled. If answered in time, the time needed is added to hint.

For displaying leaderboard next to content answering and statistics, add a tab for leaderboard route.