Extend participants series overview for quiz mode
If quiz mode is enabled, the leaderboard position, a new chart with score and leaderboard list is displayed in series overview.

Furthermore duration for the single contents is added to content list items.

For now, leaderboard is displayed initially if quiz mode is enabled and contents can be displayed additionally. In the future we could use tabs to switch between these lists.