Use tabs to switch views in participants content answer card
Instead of flipping the card, use tab navigation to switch between answering, stats and other future routes. For this `ContentResultsComponent` has been adjusted so that the already displayed rendered content body can be reused for all routes and unnecessary controls (show/hide stats, show correct) are hidden.

With this changes, styling of the `StepperComponent` can be simplified.