Fix unsubscribing on entity cache entry dispose
Websocket/STOMP subscriptions for cache entries
were not cancelled when entries were removed from
cache. This lead to errors when entity change
events were received.

Because of an identifier mismatch when comparing
the cache key type, dispose handlers were not
called. The same name, which is used for entity
cache keys, is now also used for registering the
handler.

Closes: #1323