Disable content groups for moving or copying contents if content isn't compatible with group type
Contents can either way be moved to `MIXED` groups or groups with the same group type. Contents also can be moved if content format is compatible with group type and choice or binary content has either no correct answers when moving to survey or has correct answers if moved to quiz.