Improve scale calculation for presentation mode
Instead of taking only the width into account for the scale calculation, use the aspect ration as well. This and slightly reducing the scaling lead to a more accurate fit for different screen sizes – from tablet to 4K.

Also use this calculation for scaling of live feedback component - the changes shouldn't be really noticable but ensures consistency.

For contents this changes ensure a better fit for contents with more than 5 answer options, leaderboard and attributions page.