Add foundation user-specified room content language
In addition to a the language property for rooms, this adds a directive
which sets the text direction based on the language. The text direction
is set automatically based on the locale for the language if supported
by the browser. Otherwise auto is used.

The languageDirection pipe is provided for cases, where the directive
cannot be used.

Related: #1333