Set lang and dir attributes for user-generated content
The attributes are set based on the room language property. While this
will not work for cases where multiple languages are used inside a room,
it is still better than assuming all content is in the language of the
UI.

The UI was not designed with RTL text in mind, so the dir attribute can
not be applied in all instances of user-generated content. Still, it can
be applied in the places where it matters most, including:

* room description in the overview
* content carousel
* comments

Furthermore, this sets the text direction for textareas to auto. For
text inputs it is not set because they are usually tightly coupled with
other elements which do not work well with RTL at this time.