Use component input binding instead of route snapshot
Instead of manually use data from route snapshot, just use input variables if possible. If route data isn't bind to component, it shouldn't use it anyway, so in this case pass data from parent component.

Wherever possible, `required` flag is set for inputs. If a default is needed, `transform` is used. If route data is used to set another variable `set` is used.

If routes child data is used, input binding cannot be used. But we might want to refactor this anyway in the future, since this is very hard to understand and can easily lead to bugs. See `NavBarComponent` for example where child route data is used: `this.route.children[0]?.snapshot.data.commentSettings?.disabled`.

The changes also affect stories which now can be modified more easily via input args.