Fix broken editor's content carousel for guest accounts
This fixes the request for user settings which failed with 403 due to
new restrictions introduced on the backend side.

There is no need to use the find endpoint to look up the current user
since the user ID is known. The entity is now retrieved by ID which also
simplifies the logic.

Closes: #1424