Allow private content group templates
Publishing content group templates is now optional. As default, only name, description and language are needed to create a template. If switched to publishing, license, attribution and tags are needed.

For private templates, no license is displayed in template list. Also the tag section has been removed for private ones.