Perform logout from SSO when logging out
Previously, the logout does not really do anything when SSO is used for
authentication because the authentication is still valid on the SSO, so
the client can reauthenticate automatically.

A single signout URL is now opened when logging out to globally logout
the user. This feature currently is experimental and needs to be
explicitly enabled (`ui.sso?.singleLogoutEnabled`).

Related: particify/dev/foss/arsnova-server#286
Depends-on: particify/dev/foss/arsnova-server!888