Improve room overview
Instead of having different layouts for the features in a room, use a card with the same structure for all. Cards contain the feature icon, feature name and a short description, depending on the role.

Next to feature status, for Q&A the post count and for Live Feedback the answer count is displayed.

Editors have options to start/stop Q&A and Live Feedback from the room overview.

For this move feature specific logic for Q&A and Live Feedback to feature card components.

Also remove no longer used components (RoomActionButtonComponent, RoomOverviewHeaderComponent).