Improve question series management
Improve UX for larger devices (> 999px) by displaying the content list next to a card, which, depending on the state, contain content creation/editing, results or leaderboard.

Both parts of the view are synchronized with events, so that e.g. answer count is updated if there are incoming answers at the result card or results are reset if done so in list.

For mobile devices, the old behavior with subroutes is used. For this a back button has been introduced.

To use content editing and content results in different views, the components have been adjusted. Also wherever possible, hard coded styling is replaced with SASS variables.

In content results component, correct answer is displayed by default now and chart controls have been moved to fixed fab button section.

In content editing, tabs have been introduced to switch between editing and preview.